moved type location
